lib/addoninfo.cpp: Add JSON schema validation.
For JSON schema validation in cpp code, valijson
(https://github.com/tristanpenman/valijson) is added. It supports picojson and
is actively maintained. It is used in one-header-include style, because it
seems the simplest solution, for now at least. The header is added, alongside
the LICENSE file, in externals/valijson/.

Various warnings that are emitted compiling the default code generated by
valijson are fixed; these fixes are pushed upstream.

As JSON is now validated against a schema, various ad-hoc checks could be
safely removed.

A simple way to test addon JSON schema validation, is to invoke cppcheck using:

   cppcheck --addon='{"script":1}' test.c

A minor bugfix is that the file specified by "executable" was searched using
getFullPath() with fileName as the second argument; it looks like this should
have been exename.

The fileName argument of parseAddonInfo() is only used for log messages, and is
now replaced by the string "inline JSON" instead of the actual JSON string that
was parsed. The command as given above will therefore output:

   Loading inline JSON failed. JSON schema validation failed: <root> [script]
   Value type not permitted by 'type' constraint., <root> Failed to validate
   against schema associated with property name 'script'.

The addon JSON unit tests are also updated and now operate on JSON specified on
the command line.