reduced `Tokenizer::isCPP()` usage / SymbolDatabase: removed `mIsCpp` and `isCPP()` (#5725)
Each `Token` (should) be connected to a `TokenList`. `Tokenizer` just
encapsulates that so we have no need to check the `Tokenizer` but can
simply ask the `Token`.

Also if we have function calls we pass in a flag to tell it if it is
C/C++ we can get rid of that flag and simply ask the `Token`. In several
cases we were only passing that flag to be used in some underlying calls
making it completely unnecessary in the top call.