fixed #12666 - got rid of ill-conceived  `USE_MATCHCOMPILER` CMake option validation (#6348)
I went quite overboard when I introduced the validation of the define.
That was not necessary in CMake as we provide feedback on the usage of
the matchcompiler. So only handle the special options and otherwise rely
on the regular CMake logic.

(cherry picked from commit c4be14dde2a1217d07df65c9cdd08a4b5915386e)