allow CLICOLOR_FORCE env var to force CLI color output (#6270)
This adds support for using the `CLICOLOR_FORCE` environment variable to
force color output in Linux even when a tty is not used, mimicking
behavior in [CMake and other
tools](https://bixense.com/clicolors/#bug-reports). The primary intent
is to allow color to be output when cppcheck is run in CI, for example.