Added Support for Shared Items Projects (#6286)
At work (Areus Gmbh), we use Shared Items Projects in order to
deduplicate code. In contrast to conventional project files
(`.vcxproj`), the compilation configuration is set by the "includee",
which makes the code-reuse of the project more flexible.

This pull-request adds support for Shared Items Projects (`.vcxitems`)
in Visual studio projects / solutions.

The idea is that for each `.vcxproj`, we check for referenced shared
items projects. When all the source files and include paths are build,
we add the ones defined in the referenced `.vcxitems`.
The implementation probably isn't 100% robust, but this works for our
projects and could serve as a starting point :)

I apologize for the commit history - it's a bit wonky as we currently
still use 2.11. I tried to merge the changes to the main branch so other
could benefit from these changes as well, but seems like the git history
looks a bit weird as a result.

---------

Co-authored-by: Felix Faber <felix.faber@areus.de>