Fix BOOST_REQUIRE false positives (#7341)
The **BOOST_REQUIRE** macros would not exit early and in case of a
possible null pointer assignment in the condition, it would not get
caught in the macro definition.
This would show false positives in next usages of the condition and show
a `nullPointerOutOfMemory` warning, whereas the **BOOST_REQUIRE** would
have thrown an exception.

Mimic **BOOST_REQUIRE** behavior in the macro.

Added a test for this, also the warning would not show when using
`std::malloc`, but does show up when using `malloc`.