Fix #8089 (feature request: Show date/time of check in HTML report) (#7305)
This is to fix the bug:
https://trac.cppcheck.net/ticket/8089
Title: 
feature request: Show date/time of check in HTML report
Description:
Cppcheck HTML report is a helpful tool to automate checking. It could be
extended by the date/time information, when the checking was finished.
This date/time information assists developers in fixing issues.