Misra 17.3 checks finds implicit function calls (#7327)
I am not sure why the misra check for 17.3 was reduced to while and if
statements but i think that shouldn't be the case.

Now it should detect more / all functions which are implicitly called.

I had a bit of a issue with function pointer defintion as the tokens

```
int16_t ( * misra_8_2_p_a ) ()
```
cause the detection to trigger on the int16_t. But i filtered these
false function pointer detection's explicitly.


I also updated the test with the expected errors. As C89 is used i
needed to add more expected 17.3 as needed.
* printf_s
* UINT32_C
* va_copy

are all not present in C89. But even if we bump up to C11 this PR
https://github.com/danmar/cppcheck/pull/7325 would need to be merged
before hand to actually get all the needed Identifiers.