Fix #1420 (GUI: Clarify results view context menu) (#7674)
The recheck and hide options in the context menu now show how many files
or errors are affected. If no files or errors are selected, the
pertinent option is disabled. I've opted to count top-level header files
as well since the recheck option is effective on them, they're rechecked
as part of the file that included them.