Fixup #14021 (Add file mode specifier to PathMatch) (#7689)
This is an addendum to #7645 that adds the ability to specify
directory-only matching in PathMatch patterns with a trailing slash. I
originally didn't add this because I wanted to keep PathMatch purely
syntactic, and this feature seemed like it would require filesystem
calls in the PathMatch code.

This PR solves that problem by lifting the responsibility of file mode
checking to the caller, thus keeping PathMatch purely syntactic while
still supporting directory-only matching.

Previously `/test/foo/` would match `/test/foo` even if `foo` is a
regular file. With this change the caller can specify the file mode of
the file named by the provided path, and `/test/foo` will only match
when the file mode is directory.

The semantics of patterns that do not have a trailing slash is
unchanged, e.g. `/test/foo` still matches `/test/foo/` and
`/test/foo/bar.cpp`, regardless of the file mode.

Also adds some more tests.