enabled `information` messages in selfchecks / mitigated `unmatchedSuppression` warnings (#3090)
The `unmatchedSuppression` messages might be a bug where the regular
expression appears to be incorrectly validated. I saw this with local
runs but didn't look into it yet. There was also different behavior
between using no threads and `-j`. I will file tickets as soon as I have
the time to look at it.