Make it easier to disable process executor (#7876)
By adding a DISALLOW_PROCESS_EXECUTOR build option it's possible to
conveniently disable the usage of fork() on non-Window platforms that
either don't have fork(), or have an incomplete or inefficient
implementation that is to be considered a last resort only. In the same
commit the DISALLOW_THREAD_EXECUTOR is used for conditionally compiling
the thread executor so that both executors are treated in the same way.